requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
tqdm
pymongo
qdrant-client
python-dotenv
pyyaml
